Abstract tmp_dir stuff into a separate module
