+++ multiqc/core/exec_modules.py
@@ -1,11 +1,9 @@
 import logging
-import shutil
 import sys
 import time
 import traceback
 import tracemalloc
 from importlib.metadata import EntryPoint
-from pathlib import Path
 from typing import Dict, Union, Callable, List
 
 import rich
@@ -13,9 +11,8 @@ from rich.syntax import Syntax
 
 from multiqc import config, report
 from multiqc.base_module import BaseMultiqcModule, ModuleNoSamplesFound
+from multiqc.core import plugin_hooks, software_versions, tmp_dir
 from multiqc.core.exceptions import RunError
-from multiqc.core import plugin_hooks, software_versions
-from multiqc.validation import ConfigValidationError
 from multiqc.modules.software_versions import MultiqcModule as SoftwareVersionModule
 
 logger = logging.getLogger(__name__)
@@ -137,9 +134,9 @@ def exec_modules(
                     panel_width = max(tb_width, log_width)
                     return rich.console.Measurement(panel_width, panel_width)
 
-            from multiqc.core.init_log import rich_console
+            from multiqc.core.log_and_rich import rich_console_print
 
-            rich_console.print(
+            rich_console_print(
                 rich.panel.Panel(
                     CustomTraceback(),
                     title=f"Oops! The '[underline]{this_module}[/]' MultiQC module broke...",
@@ -188,8 +185,8 @@ def exec_modules(
     # Did we find anything?
     if len(report.modules) == 0:
         logger.warning("No analysis results found. Cleaning up…")
-        if clean_up and report.tmp_dir and Path(report.tmp_dir).exists():
-            shutil.rmtree(report.tmp_dir)
+        if clean_up:
+            tmp_dir.clean_up()
         logger.info("MultiQC complete")
         # Exit with an error code if a module broke
         raise RunError(sys_exit_code=sys_exit_code)
