+++ multiqc/core/log_and_rich.py
@@ -2,12 +2,13 @@
 Code to initialise the MultiQC logging
 """
 
+import datetime
 import logging
 import os
 import shutil
 import sys
-import tempfile
 from pathlib import Path
+from typing import List, Optional
 
 import coloredlogs  # type: ignore
 import rich
@@ -15,15 +16,17 @@ from rich.logging import RichHandler
 from rich.theme import Theme
 import rich.jupyter
 import rich_click
+import rich.progress
+from tqdm import tqdm
 
 from multiqc import config
+from multiqc.core import tmp_dir
 from multiqc.utils import util_functions
+from multiqc.utils.util_functions import is_running_in_notebook
 
+log_tmp_fn = Path("/dev/null")
 
-log_tmp_dir = None
-log_tmp_fn = "/dev/null"
-
-rich_console: rich.console.Console
+rich_console: Optional[rich.console.Console] = None
 
 
 def init_log():
@@ -35,11 +38,10 @@ def init_log():
 
     loglevel (str): Determines the level of the log output.
     """
-    global log_tmp_dir, log_tmp_fn
+    global log_tmp_fn
     # Have to create a separate directory for the log file otherwise Windows will complain
     # about same file used by different processes:
-    log_tmp_dir = tempfile.mkdtemp()
-    log_tmp_fn = os.path.join(log_tmp_dir, "multiqc.log")
+    log_tmp_fn = tmp_dir.get_tmp_dir() / "multiqc.log"
 
     logger = logging.getLogger()  # root logger
 
@@ -51,6 +53,9 @@ def init_log():
     if config.quiet:
         log_level = "WARNING"
     logger.setLevel(log_level)
+
+    # Remove DEBUG level for the PIL.PngImagePlugin logger
+    logging.getLogger("PIL").setLevel(logging.INFO)
 
     # Automatically set no_ansi if not a tty terminal
     if config.no_ansi is False:
@@ -100,6 +105,9 @@ def init_log():
     file_handler.setFormatter(logging.Formatter(debug_template))
     logger.addHandler(file_handler)
 
+    logger.debug(f"Using temporary directory: {tmp_dir}")
+    logger.debug(f"Logging to file: {log_tmp_fn}")
+
 
 def _setup_coloredlogs(log_level, logger, debug_template):
     # Use coloredlogs as Rich is breaking output formatting
@@ -135,11 +143,11 @@ def _setup_coloredlogs(log_level, logger, debug_template):
     logger.propagate = False
 
 
-def _setup_rich_handler(log_level, logger):
+def _setup_rich_handler(log_level, console, logger):
     # Set up the console logging stream
     console_handler = RichHandler(
         level=log_level,
-        console=rich_console,
+        console=console,
         show_level=log_level == "DEBUG",
         show_time=log_level == "DEBUG",
         log_time_format="[%Y-%m-%d %H:%M:%S]",
@@ -192,7 +200,7 @@ def _print_intro_with_coloredlogs():
         DIM = ""
         DARK_ORANGE = ""
         RESET = ""
-    emoji = util_functions.choose_emoji()
+    emoji = choose_emoji()
     emoji = f" {emoji}" if emoji else ""
     intro = f"{DARK_ORANGE}///{RESET} {BOLD}https://multiqc.info{RESET}{emoji} {DIM}v{config.version}{RESET}"
     if not util_functions.is_running_in_notebook():
@@ -201,7 +209,8 @@ def _print_intro_with_coloredlogs():
 
 
 def _print_intro_with_rich():
-    rich_console.print(f"\n{rich_click.rich_click.HEADER_TEXT}\n")
+    if rich_console is not None:
+        rich_console.print(f"\n{rich_click.rich_click.HEADER_TEXT}\n")
 
 
 def move_tmp_log():
@@ -212,13 +221,12 @@ def move_tmp_log():
     # https://stackoverflow.com/questions/15435652/python-does-not-release-filehandles-to-logfile
     logging.shutdown()
 
-    if config.data_dir is None or not Path(config.data_dir).is_dir() or not os.path.exists(log_tmp_fn):
+    if config.data_dir is None or not Path(config.data_dir).is_dir() or not log_tmp_fn.exists():
         return
 
     try:
         shutil.copy(log_tmp_fn, os.path.join(config.data_dir, "multiqc.log"))
         os.remove(log_tmp_fn)
-        util_functions.rmtree_with_retries(log_tmp_dir)
     except IOError:
         pass
 
@@ -253,3 +261,124 @@ def force_term_colors():
     if os.getenv("GITHUB_ACTIONS") or os.getenv("FORCE_COLOR") or os.getenv("PY_COLORS"):
         return True
     return None
+
+
+emoji_rich_ids = {
+    "🍾": ":bottle_with_popping_cork:",
+    "🌹": ":rose:",
+    "🍀": ":four_leaf_clover:",
+    "🌏": ":globe_showing_asia-australia:",
+    "🎃": ":jack-o-lantern:",
+    "🎅": ":santa:",
+    "🎄": ":christmas_tree:",
+    "🔍": ":mag:",
+}
+
+emoji_dates = {
+    "🍾": (1, 1, 1, 5),  # New Year's Day
+    "🌹": (2, 14, 0, 0),  # Valentine's Day
+    "🍀": (3, 17, 0, 0),  # St Patrick's Day
+    "🌏": (4, 22, 0, 0),  # Earth Day
+    "🎃": (10, 31, 5, 0),  # Halloween
+    "🎅": (12, 25, 0, 0),  # Christmas Day
+    "🎄": (12, 25, 7, 7),  # Christmas
+}
+
+
+def choose_emoji(use_rich=False) -> str:
+    """Choose an emoji to use in the report header."""
+    # NB: We haven't parsed the config yet, so can't disable via config
+    if _no_unicode():
+        return ""
+
+    today = datetime.date.today()
+
+    selected_emoji = "🔍"
+    for emoji, (month, day, days_before, days_after) in emoji_dates.items():
+        special_date = datetime.date(today.year, month, day)
+        date_range_start = special_date - datetime.timedelta(days=days_before)
+        date_range_end = special_date + datetime.timedelta(days=days_after)
+        if date_range_start <= today <= date_range_end:
+            selected_emoji = emoji
+    if use_rich:
+        return emoji_rich_ids[selected_emoji]
+    return selected_emoji
+
+
+def _no_unicode() -> bool:
+    # When LANG or PYTHONIOENCODING or is not set, Rich won't be able to print fancy unicode
+    # characters for the progress bar, and the runtime would crash with UnicodeEncodeError:
+    # https://github.com/MultiQC/MultiQC/actions/runs/8814275065/job/24193771822
+    # See https://github.com/Textualize/rich/issues/212
+    return (
+        "utf".casefold() not in os.environ.get("LANG", "").casefold()
+        and "utf".casefold() not in os.environ.get("PYTHONIOENCODING", "").casefold()
+    )
+
+
+def iterate_using_progress_bar(items: List, update_fn, item_to_str_fn, desc, disable_progress=False):
+    # GitHub actions doesn't understand ansi control codes to move the cursor,
+    # so it prints each update ona a new line. Better disable it for CI.
+    disable_progress = disable_progress or config.no_ansi or config.quiet or os.getenv("CI") is not None
+
+    # Rich widgets do not look good in Jupyter, of it there is no unicode support.
+    # Additionally, falling back to tqdm if rich_console was not initialized. That
+    # happens when init_log.init_log() wasn't run, i.e in unit tests.
+    if is_running_in_notebook() or _no_unicode() or rich_console is None:
+        # ANSI escape code for dim text
+        if not config.no_ansi:
+            DIM = "\033[2m"
+            BLUE = "\033[34m"
+            RESET = "\033[0m"
+        else:
+            DIM = ""
+            BLUE = ""
+            RESET = ""
+
+        bar_format = f"{BLUE}{desc:>17} {RESET}| " + "{bar:40} {percentage:3.0f}% {n_fmt}/{total_fmt} {desc}"
+
+        # Set up the tqdm progress bar
+        with tqdm(
+            total=len(items),
+            desc=desc,
+            unit="file",
+            file=sys.stdout,
+            disable=disable_progress,
+            bar_format=bar_format,
+        ) as pbar:
+            for i, item in enumerate(items):
+                pbar.update(1)
+                # Update the progress bar description with the file being searched
+                pbar.set_description_str(f"{DIM}{item_to_str_fn(item)[-50:]}{RESET}")
+                update_fn(i, item)
+
+            # Clear the description after the loop is complete
+            pbar.set_description("")
+            pbar.refresh()
+    else:
+        N_SPACES_BEFORE_PIPE = 15  # to align bar desc with other log entries
+        need_to_add_spaces = max(0, N_SPACES_BEFORE_PIPE - len(desc))
+        progress_obj = rich.progress.Progress(
+            "[blue][/]" + " " * need_to_add_spaces,
+            rich.progress.SpinnerColumn(),
+            "[blue]{task.description}[/] |",
+            rich.progress.BarColumn(),
+            "[progress.percentage]{task.percentage:>3.0f}%",
+            "[green]{task.completed}/{task.total}",
+            "[dim]{task.fields[s_fn]}[/]",
+            console=rich_console,
+            disable=disable_progress,
+        )
+        with progress_obj as progress:
+            mqc_task = progress.add_task(desc, total=len(items), s_fn="")
+            for i, item in enumerate(items):
+                progress.update(mqc_task, advance=1, s_fn=item_to_str_fn(item)[-50:])
+                update_fn(i, item)
+            progress.update(mqc_task, s_fn="")
+
+
+def rich_console_print(*args, **kwargs):
+    if rich_console is not None:
+        rich_console.print(*args, **kwargs)
+    else:
+        print(*args, **kwargs)
