+++ multiqc/utils/util_functions.py
@@ -1,52 +1,15 @@
 """MultiQC Utility functions, used in a variety of places."""
 
-from typing import Dict, List
-
-import array
+import datetime
 import json
 import logging
 from collections import defaultdict, OrderedDict
+from typing import Dict
 
-import os
-import shutil
-import sys
-import time
-import datetime
+import array
 import math
 
-import rich
-import rich.progress
-from tqdm import tqdm
-
-from multiqc import config
-from multiqc.core import init_log
-
 logger = logging.getLogger(__name__)
-
-
-def rmtree_with_retries(path, _logger=None, max_retries=10):
-    """
-    Robustly tries to delete paths.
-    Retries several times (with increasing delays) if an OSError
-    occurs.  If the final attempt fails, the Exception is propagated
-    to the caller.
-    """
-
-    for i in range(max_retries):
-        try:
-            shutil.rmtree(path)
-            return
-        except OSError:
-            if _logger:
-                _logger.info(f"Unable to remove path: {path}")
-                _logger.info(f"Retrying after {i**2} seconds")
-            else:
-                print(f"Unable to remove path: {path}", file=sys.stderr)
-                print(f"Retrying after {i**2} seconds", file=sys.stderr)
-            time.sleep(i**2)
-
-    # Final attempt, pass any Exceptions up to caller.
-    shutil.rmtree(path)
 
 
 def strtobool(val) -> bool:
@@ -69,48 +32,6 @@ def strtobool(val) -> bool:
         return False
     else:
         raise ValueError(f"invalid truth value {val!r}")
-
-
-emoji_rich_ids = {
-    "🍾": ":bottle_with_popping_cork:",
-    "🌹": ":rose:",
-    "🍀": ":four_leaf_clover:",
-    "🌏": ":globe_showing_asia-australia:",
-    "🎃": ":jack-o-lantern:",
-    "🎅": ":santa:",
-    "🎄": ":christmas_tree:",
-    "🔍": ":mag:",
-}
-
-emoji_dates = {
-    "🍾": (1, 1, 1, 5),  # New Year's Day
-    "🌹": (2, 14, 0, 0),  # Valentine's Day
-    "🍀": (3, 17, 0, 0),  # St Patrick's Day
-    "🌏": (4, 22, 0, 0),  # Earth Day
-    "🎃": (10, 31, 5, 0),  # Halloween
-    "🎅": (12, 25, 0, 0),  # Christmas Day
-    "🎄": (12, 25, 7, 7),  # Christmas
-}
-
-
-def choose_emoji(rich=False) -> str:
-    """Choose an emoji to use in the report header."""
-    # NB: We haven't parsed the config yet, so can't disable via config
-    if no_unicode():
-        return ""
-
-    today = datetime.date.today()
-
-    selected_emoji = "🔍"
-    for emoji, (month, day, days_before, days_after) in emoji_dates.items():
-        special_date = datetime.date(today.year, month, day)
-        date_range_start = special_date - datetime.timedelta(days=days_before)
-        date_range_end = special_date + datetime.timedelta(days=days_after)
-        if date_range_start <= today <= date_range_end:
-            selected_emoji = emoji
-    if rich:
-        return emoji_rich_ids[selected_emoji]
-    return selected_emoji
 
 
 def replace_defaultdicts(data):
@@ -199,17 +120,6 @@ def is_running_in_notebook() -> bool:
     return False
 
 
-def no_unicode() -> bool:
-    # When LANG or PYTHONIOENCODING or is not set, Rich won't be able to print fancy unicode
-    # characters for the progress bar, and the runtime would crash with UnicodeEncodeError:
-    # https://github.com/MultiQC/MultiQC/actions/runs/8814275065/job/24193771822
-    # See https://github.com/Textualize/rich/issues/212
-    return (
-        "utf".casefold() not in os.environ.get("LANG", "").casefold()
-        and "utf".casefold() not in os.environ.get("PYTHONIOENCODING", "").casefold()
-    )
-
-
 def compress_number_lists_for_json(obj):
     """
     Take an object that should be JSON and compress all the lists of integer
@@ -284,64 +194,3 @@ def update_dict(target: Dict, source: Dict, none_only=False):
                 else:
                     target[key] = src_val
     return target
-
-
-def iterate_using_progress_bar(items: List, update_fn, item_to_str_fn, desc, disable_progress=False):
-    # GitHub actions doesn't understand ansi control codes to move the cursor,
-    # so it prints each update ona a new line. Better disable it for CI.
-    disable_progress = disable_progress or config.no_ansi or config.quiet or os.getenv("CI") is not None
-
-    # Rich widgets do not look good in Jupyter, of it there is no unicode support.
-    # Additionally, falling back to tqdm if rich_console was not initialized. That
-    # happens when init_log.init_log() wasn't run, i.e in unit tests.
-    if is_running_in_notebook() or no_unicode() or not getattr(init_log, "rich_console", None):
-        # ANSI escape code for dim text
-        if not config.no_ansi:
-            DIM = "\033[2m"
-            BLUE = "\033[34m"
-            RESET = "\033[0m"
-        else:
-            DIM = ""
-            BLUE = ""
-            RESET = ""
-
-        bar_format = f"{BLUE}{desc:>17} {RESET}| " + "{bar:40} {percentage:3.0f}% {n_fmt}/{total_fmt} {desc}"
-
-        # Set up the tqdm progress bar
-        with tqdm(
-            total=len(items),
-            desc=desc,
-            unit="file",
-            file=sys.stdout,
-            disable=disable_progress,
-            bar_format=bar_format,
-        ) as pbar:
-            for i, item in enumerate(items):
-                pbar.update(1)
-                # Update the progress bar description with the file being searched
-                pbar.set_description_str(f"{DIM}{item_to_str_fn(item)[-50:]}{RESET}")
-                update_fn(i, item)
-
-            # Clear the description after the loop is complete
-            pbar.set_description("")
-            pbar.refresh()
-    else:
-        N_SPACES_BEFORE_PIPE = 15  # to align bar desc with other log entries
-        need_to_add_spaces = max(0, N_SPACES_BEFORE_PIPE - len(desc))
-        progress_obj = rich.progress.Progress(
-            "[blue][/]" + " " * need_to_add_spaces,
-            rich.progress.SpinnerColumn(),
-            "[blue]{task.description}[/] |",
-            rich.progress.BarColumn(),
-            "[progress.percentage]{task.percentage:>3.0f}%",
-            "[green]{task.completed}/{task.total}",
-            "[dim]{task.fields[s_fn]}[/]",
-            console=init_log.rich_console,
-            disable=disable_progress,
-        )
-        with progress_obj as progress:
-            mqc_task = progress.add_task(desc, total=len(items), s_fn="")
-            for i, item in enumerate(items):
-                progress.update(mqc_task, advance=1, s_fn=item_to_str_fn(item)[-50:])
-                update_fn(i, item)
-            progress.update(mqc_task, s_fn="")
